fix(gl-react): glsl error resilient & shader cache + test
glsledit used to crash if you were to write a wrong shader (which is at any new keystroke basically) and THEN if you come back to an old shader.
we were still using the old shader object even tho it was released.
this removes a wrong assumption that was taken in <Node> that Node hold the shader object, it's no longer true as it comes from the <Surface> and it caches there,
therefore, the <Node> must not .dispose() the shader object

the issue now however is that glsledit example will start leaking shader objects as we cache them, we will fix that using a LRU